Add reachability table for randomizer logic
The table is not complete yet, but should eventually contain a matrix describing the connections between each room exit and pickup (including powerup requirements). The can later be used to check if a seed is possible or not. Furthermore if we figure out room randomization, we can simply let the randomizer shuffle the connections between rooms.