Calculate reduced connection matrix
I added a bunch of functions to work with the connection matrix and compress it into a format the randomizer can use for generating a map. Since we're only interested in which pickup can be reached from which other pickup (to automatically check if a set of powerup locations can be solved or not). A reduced connection matrix with only the pickup locations was added.

At the moment the matrix reduction code is way too slow to be used for a full room randomizer, but for everything else we can precalculate the matrices needed to generate maps.

Also made some corrections in the full matrix