Fixed bug in requirement matrix
Using the new debug tool I found the problem in the matrix which allowed a connection from the bottom area back to the spawn are using only gloves.
The problem was in Abstract Bridge where a connection existed from the left entrance to the bottom and from the bottom to the right side with only gloves. Actually crossing to the other side is impossible with just gloves though, so I split the bottom screen edge up into a right and left side.